Added: Subject Schedule and Subject Fee cache
