+++ app/src/main/java/io/zoemeow/dutapp/MainActivity.kt
@@ -126,7 +126,7 @@ fun NavigationHost(
 ) {
     NavHost(
         navController = navController,
-        startDestination = NavRoutes.News.route,
+        startDestination = NavRoutes.Settings.route,
         modifier = Modifier.padding(padding)
     ) {
         composable(NavRoutes.Home.route) {
