+++ app/src/main/java/io/zoemeow/dutapp/view/Settings.kt
@@ -1,31 +1,16 @@
 package io.zoemeow.dutapp.view
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
+import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.platform.LocalFocusManager
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.unit.dp
-import io.zoemeow.dutapp.R
-import io.zoemeow.dutapp.model.AccountInformationItem
+import androidx.compose.ui.window.DialogProperties
 import io.zoemeow.dutapp.viewmodel.MainViewModel
 
 @Composable
@@ -46,277 +31,18 @@ fun Settings(mainViewModel: MainViewModel) {
     }
 }
 
+@OptIn(ExperimentalComposeUiApi::class)
 @Composable
 fun SettingsAccountTag(mainViewModel: MainViewModel) {
-    Column(modifier = Modifier
-        .fillMaxSize()
-        .padding(20.dp)) {
-        // Login
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clip(RoundedCornerShape(10.dp))
-                .background(MaterialTheme.colorScheme.secondaryContainer)
-                .border(2.dp, MaterialTheme.colorScheme.secondary, RoundedCornerShape(10.dp))
-                .clickable {
-                    if (mainViewModel.accCacheData.value.sessionID.value.isEmpty() && !mainViewModel.isAvailableOffline())
-                        mainViewModel.accountPaneIndex.value = 1
-                    else mainViewModel.accountPaneIndex.value = 3
-                }
-        ) {
-            if (mainViewModel.accCacheData.value.sessionID.value.isEmpty() && !mainViewModel.isAvailableOffline()) AccountTagNotLoggedIn()
-            else AccountTagLoggedIn(id = mainViewModel.accCacheData.value.accountInformationData.value.studentId ?: String())
-        }
-    }
-}
-
-@Composable
-fun AccountTagNotLoggedIn() {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(15.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        Text(
-            text = stringResource(id = R.string.navlogin_screennotloggedin_text1),
-            style = MaterialTheme.typography.headlineSmall
-        )
-        Spacer(modifier = Modifier.size(5.dp))
-        Text(
-            text = stringResource(id = R.string.navlogin_screennotloggedin_text2),
-            style = MaterialTheme.typography.bodyLarge
-        )
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun AccountPageLogin(
-    mainViewModel: MainViewModel,
-    backRequest: () -> Unit,
-) {
-    val user = remember { mutableStateOf(String()) }
-    val pass = remember { mutableStateOf(String()) }
-    val autoLogin = remember { mutableStateOf(false) }
-    val passTextFieldFocusRequester = remember { FocusRequester() }
-    val focusManager = LocalFocusManager.current
-
-    val loginCommand: () -> Unit = {
-        focusManager.clearFocus()
-        mainViewModel.login(user.value, pass.value, autoLogin.value)
-        pass.value = String()
-    }
-
-    Scaffold(
-        content = { innerPadding ->
-            Box(modifier = Modifier.padding(innerPadding)) {
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(20.dp)
-                        .navigationBarsPadding()
-                        .wrapContentHeight(),
-                    verticalArrangement = Arrangement.Top,
-                ) {
-                    Text(
-                        text = stringResource(id = R.string.navlogin_screenlogin_title),
-                        style = MaterialTheme.typography.headlineMedium
-                    )
-                    OutlinedTextField(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(top = 10.dp),
-                        value = user.value,
-                        label = { Text(stringResource(id = R.string.navlogin_screenlogin_username)) },
-                        keyboardOptions = KeyboardOptions(imeAction = ImeAction.Next),
-                        keyboardActions = KeyboardActions(
-                            onGo = { passTextFieldFocusRequester.requestFocus() }
-                        ),
-                        onValueChange = { user.value = it },
-                    )
-                    OutlinedTextField(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(top = 10.dp)
-                            .focusRequester(passTextFieldFocusRequester),
-                        value = pass.value,
-                        label = { Text(stringResource(id = R.string.navlogin_screenlogin_password)) },
-                        visualTransformation = PasswordVisualTransformation(),
-                        keyboardOptions = KeyboardOptions(
-                            keyboardType = KeyboardType.Password,
-                            imeAction = ImeAction.Go
-                        ),
-                        keyboardActions = KeyboardActions(onGo = { loginCommand() }),
-                        onValueChange = { pass.value = it },
-                    )
-                    // Check box: Auto login
-                    Row(
-                        modifier = Modifier.clickable { autoLogin.value = !autoLogin.value },
-                        verticalAlignment = Alignment.CenterVertically,
-                    ) {
-                        Checkbox(
-                            checked = autoLogin.value,
-                            onCheckedChange = { autoLogin.value = it },
-                        )
-                        Spacer(modifier = Modifier.size(10.dp))
-                        Text(
-                            text = stringResource(id = R.string.navlogin_screenlogin_rememberlogin)
-                        )
-                    }
-                    Row(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(top = 10.dp),
-                        horizontalArrangement = Arrangement.Center
-                    ) {
-                        Button(
-                            onClick = loginCommand,
-                            content = { Text(stringResource(id = R.string.navlogin_screenlogin_btnlogin)) },
-                        )
-                        Spacer(modifier = Modifier.size(10.dp))
-                        Button(
-                            onClick = backRequest,
-                            content = { Text(stringResource(id = R.string.navlogin_screenlogin_btncancel)) },
-                        )
-                    }
-                }
-            }
-        }
-    )
-}
-
-@Composable
-fun AccountTagLoggedIn(id: String) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(15.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        Text(
-            text = "ID: $id",
-            style = MaterialTheme.typography.headlineSmall
-        )
-        Spacer(modifier = Modifier.size(5.dp))
-        Text(
-            text = "Tap here to view your info/logout/re-login your account.",
-            style = MaterialTheme.typography.bodyLarge
-        )
-    }
-}
-
-@Composable
-fun AccountPageLoggingIn() {
-    Surface(modifier = Modifier.fillMaxSize()) {
-        Column(
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Text(
-                text = "Please wait",
-                style = MaterialTheme.typography.headlineSmall
-            )
-            Spacer(modifier = Modifier.size(5.dp))
-            Text(
-                text = "Logging you in...",
-                style = MaterialTheme.typography.bodyLarge
-            )
-            Spacer(modifier = Modifier.size(15.dp))
-            LinearProgressIndicator(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(start = 90.dp, end = 90.dp)
-            )
-        }
-    }
-}
-
-@Composable
-fun AccountPageInformation(accInfo: AccountInformationItem, isLoading: Boolean, logout: () -> Unit, reLogin: () -> Unit) {
     val openDialog = remember { mutableStateOf(false) }
 
-    if (isLoading) AccountPageLoadingYourInfo()
-    else {
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Top
-        ) {
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(150.dp)
-                    .clip(RoundedCornerShape(bottomStart = 50.dp, bottomEnd = 50.dp))
-                    .background(MaterialTheme.colorScheme.secondaryContainer)
-            ) {
-                Column(
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    verticalArrangement = Arrangement.Center,
-                    modifier = Modifier.fillMaxSize()
-                ) {
-                    Text(
-                        text = accInfo.name ?: String(),
-                        style = MaterialTheme.typography.headlineMedium
-                    )
-                    Spacer(modifier = Modifier.size(5.dp))
-                    Text(
-                        text = accInfo.studentId ?: String(),
-                        style = MaterialTheme.typography.titleMedium
-                    )
-                    Text(
-                        text = accInfo.specialization ?: String(),
-                        style = MaterialTheme.typography.titleMedium
-                    )
-                }
-            }
-
-            data class AccInfoItem(
-                val key: String,
-                val value: String,
-            )
-            val list: ArrayList<AccInfoItem> = arrayListOf(
-                AccInfoItem("Class", accInfo.schoolClass ?: String()),
-                AccInfoItem("School Email", accInfo.schoolEmail ?: String()),
-                AccInfoItem("Email", accInfo.personalEmail ?: String()),
-                AccInfoItem("Facebook URL", accInfo.facebookUrl ?: String()),
-                AccInfoItem("Phone Number", accInfo.phoneNumber ?: String()),
-            )
-
-            Spacer(modifier = Modifier.size(20.dp))
-            LazyColumn {
-                items(list) { item ->
-                    Column(modifier = Modifier.padding(top = 8.dp, bottom = 8.dp, start = 10.dp, end = 10.dp)) {
-                        Text(
-                            text = item.key,
-                            style = MaterialTheme.typography.labelMedium
-                        )
-                        Spacer(modifier = Modifier.size(3.dp))
-                        Text(
-                            text = item.value,
-                            style = MaterialTheme.typography.bodyLarge
-                        )
-                    }
-                }
-            }
-
-            Spacer(modifier = Modifier.size(20.dp))
-            Row() {
-                Button(onClick = reLogin) {
-                    Text("Re-login")
-                }
-                Spacer(modifier = Modifier.size(5.dp))
-                Button(onClick = { openDialog.value = true }) {
-                    Text(stringResource(id = R.string.navlogin_loggedin_btnlogout))
-                }
-            }
-        }
-    }
-
+    // Alert dialog for logout
     if (openDialog.value) {
         AlertDialog(
-            modifier = Modifier.padding(15.dp),
+            properties = DialogProperties(
+                usePlatformDefaultWidth = false
+            ),
+            modifier = Modifier.padding(15.dp).fillMaxWidth().wrapContentHeight(),
             onDismissRequest = { openDialog.value = false },
             title = { Text("Logout") },
             text = {
@@ -328,7 +54,7 @@ fun AccountPageInformation(accInfo: AccountInformationItem, isLoading: Boolean, 
                 TextButton(
                     onClick = {
                         openDialog.value = false
-                        logout()
+                        mainViewModel.logout()
                     },
                     content = {
                         Text("Yes, log me out")
@@ -345,30 +71,109 @@ fun AccountPageInformation(accInfo: AccountInformationItem, isLoading: Boolean, 
             }
         )
     }
+
+    Column(modifier = Modifier.fillMaxSize()) {
+        SettingsOptionLogin(
+            mainViewModel,
+            toggleLogout = { openDialog.value = true }
+        )
+    }
 }
 
 @Composable
-fun AccountPageLoadingYourInfo() {
-    Surface(modifier = Modifier.fillMaxSize()) {
-        Column(
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Text(
-                text = "Please wait",
-                style = MaterialTheme.typography.headlineSmall
-            )
-            Spacer(modifier = Modifier.size(5.dp))
-            Text(
-                text = "Loading your information...",
-                style = MaterialTheme.typography.bodyLarge
-            )
-            Spacer(modifier = Modifier.size(15.dp))
-            LinearProgressIndicator(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(start = 90.dp, end = 90.dp)
-            )
+fun SettingsOptionLogin(mainViewModel: MainViewModel, toggleLogout: () -> Unit) {
+    Box {
+        Column {
+            Box(
+                modifier = Modifier.fillMaxWidth().height(40.dp),
+                contentAlignment = Alignment.BottomStart
+            ) {
+                Text(
+                    modifier = Modifier.padding(start = 20.dp, end = 20.dp),
+                    text = "Account",
+                    style = MaterialTheme.typography.labelLarge
+                )
+            }
+            // If not logged in
+            if (mainViewModel.accCacheData.value.sessionID.value.isEmpty() &&
+                !mainViewModel.isAvailableOffline()) {
+                // Login
+                Box(
+                    modifier = Modifier.fillMaxWidth().height(70.dp)
+                        .clickable { mainViewModel.accountPaneIndex.value = 1 },
+                    contentAlignment = Alignment.CenterStart
+                ) {
+                    Column(
+                        modifier = Modifier.padding(start = 20.dp, end = 20.dp),
+                        verticalArrangement = Arrangement.Center,
+                    ) {
+                        Text(
+                            text = "Login",
+                            style = MaterialTheme.typography.bodyLarge,
+                        )
+                        Text(
+                            text = "To use more app features, you need to sign in.",
+                            style = MaterialTheme.typography.bodySmall,
+                        )
+                    }
+                }
+            }
+            // If logged in/offline mode and logged in previously.
+            else {
+                Column {
+                    // Account Information
+                    Box(
+                        modifier = Modifier.fillMaxWidth().height(70.dp)
+                            .clickable { mainViewModel.accountPaneIndex.value = 3 },
+                        contentAlignment = Alignment.CenterStart,
+                    ) {
+                        Text(
+                            modifier = Modifier.padding(start = 20.dp, end = 20.dp),
+                            text = "View Account Information",
+                            style = MaterialTheme.typography.bodyLarge,
+                        )
+                    }
+                    // Re-login
+                    if (mainViewModel.accCacheData.value.sessionID.value.isEmpty()) {
+                        Box(
+                            modifier = Modifier.fillMaxWidth().height(70.dp)
+                                .clickable { mainViewModel.executeAutoLogin() },
+                            contentAlignment = Alignment.CenterStart,
+                        ) {
+                            Column(
+                                modifier = Modifier.padding(start = 20.dp, end = 20.dp)
+                            ) {
+                                Text(
+                                    text = "Re-login",
+                                    style = MaterialTheme.typography.bodyLarge,
+                                )
+                                Text(
+                                    text = "You have saved your login, but you haven't been logged in to system. Click here to re-login.",
+                                    style = MaterialTheme.typography.bodySmall,
+                                )
+                            }
+                        }
+                    }
+                    // Logout
+                    Box(
+                        modifier = Modifier.fillMaxWidth().height(70.dp)
+                            .clickable { toggleLogout() },
+                        contentAlignment = Alignment.CenterStart,
+                    ) {
+                        Column(modifier = Modifier.padding(start = 20.dp, end = 20.dp)) {
+                            Text(
+                                text = "Logout",
+                                style = MaterialTheme.typography.bodyLarge,
+                            )
+                            Text(
+                                text = "Logged in as ${mainViewModel.accCacheData.value.accountInformationData.value.studentId}",
+                                style = MaterialTheme.typography.bodySmall,
+                            )
+                        }
+                    }
+                }
+            }
         }
     }
 }
+
